[TASK] Respect alt text in backend thumbnails
The thumbCode method in the BackendUtility sets
the filename as the alt text for images. That's
why the backend representation of content elements
like textmedia doesn't respect any set alternative
text, whether it's set in the file_reference or
the sys_file record itself.

This change uses the alternative text and only
sets the filename as a fallback if there's no
alternative text provided for the file.

Resolves: #103285
Releases: main, 12.4
Change-Id: I3c11bb7a56f001e37eafdd1f0a998cb51fb2b460
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83232
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>