[TASK] Replace GeneralUtility::hmac usage in ext:backend
With #102761, the new `HashService` has been introduced to ensure
that HMAC generation will always use an additional secret. As a
follow-up patch, `GeneralUtility::hmac` will be deprecated when all
usages in the core have been replaced with the new `HashService`.

With this change, usages of `GeneralUtility::hmac` in ext:backend
are replaced by the new HashService.

Resolves: #103247
Related: #102761
Related: #103245
Releases: main
Change-Id: I99abec6d68397bd6c9623a9a3de776b0f6cd4a8a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83253
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>