[TASK] Migrate "move page" wizard to a modal
The "Move page" wizard is still based on the dusted `PagePositionMap`.
In order to modernize the clunky UI, the wizard is migrated to a modal
window that offers a drastically slimmed down interface, offering a page
tree to make the selection of the target page easier.

It is always possible to paste the to-be-moved page above and below a
target page, or inside a page if there are no children, yet.

Resolves: #103344
Releases: main
Change-Id: I745abb7b34e318275ae95abb689a6024e47b166e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83269
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>