[TASK] Implement `@typescript-eslint/no-array-delete` rule
The eslint configuration now takes the rule
`@typescript-eslint/no-array-delete` [1] into account, prohibiting
element removal from an array via `delete`. The problem with `delete` is
that the element to be removed is in fact replaced with an `empty` slot,
effectible keeping the array's size. Using `splice()` [2] is the
preferred way to remove elements from an array.

[1] https://typescript-eslint.io/rules/no-array-delete/
[2] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/splice

Resolves: #103375
Releases: main, 12.4
Change-Id: Ic742a43d44cbd8c5cc323f629cc82ad606a9858f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83443
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>