[TASK] Set empty alt text for file reference thumbnails
The thumbnail images in the inline file relations have the properties
alt and title set to "id=x" which is only helpful when used to show the
record uid in the tooltip. For screenreader users the info in the alt
attribute isn't helpful at all.
That's why the alt text should be empty so the thumbnail doesn't get
read out by screenreaders at all.

Resolves: #103397
Releases: main, 12.4
Change-Id: I233a54705c9c81e31b8852a8454165593fa8e061
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83431
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>