[FEATURE] Group-based content fetching in Frontend
This change adds a new RecordCollector which is similar
to CONTENT cObject, except that it does not render the records
directly, but instead creates instances of a generic Record
object, so the Templating Engine can deal with the way of
how the record should be rendered.

Based on the Page Layout / Backend Layout, all registered
colPos contents are fetched, and grouped, if the Backend Layout
has an "identifier" set via a new PageContentFetchingProcessor
("page-content").

The slide functionality can now be defined in the Page Layout as well
(Backend Layout), with a Enum defining to Collect, CollectReverse
and a simple Slide logic.

This feature brings a clearer separation of fetching
content from the database and the actual representation
in the output (fluid) in place, allowing for further
optimizations.

In addition, one needs to write much less TypoScript
to fetch all contents from a page.

Next steps is then to enrich the records with their
relations (see next feature), and to optimize the Fluid-based
rendering for any kind of content.

Resolves: #103894
Releases: main
Change-Id: I6a1a0efacefe8c83e86b9551b00199c93e284a28
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83638
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>