[TASK] Do not render `typo3-formengine-updater` if linked field is hidden
TYPO3's TCA type field is rendered in the FormEngine along with a
`<typo3-formengine-updater>` element that listens on changes on the
linked type field.

This is a rare case: if FormEngine is programmatically rendered via an
own controller and the value of the `type` field is hardcoded via
`overrideValues` in the FormEngine compiler, the field itself is not rendered.

However, the `<typo3-formengine-updater>` element is still there,
leaving an "empty" field group.

This patch now checks whether something is available for rendering
before rendering the `<typo3-formengine-updater>` element.

Resolves: #103449
Releases: main
Change-Id: I708b49665dfaa5b836954e40e99cbcb9ed44ec03
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83554
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>