[BUGFIX] Link root page title in site configuration module
To be consistent with the other listings in the
backend is the root page title of a site in the
site configuration module now linked, allowing
to edit the corresponding site configuration.

Additionally does the link as well as the
edit button now use a dedicated label instead
of a generic one.

Resolves: #103695
Releases: main, 12.4
Change-Id: Ice46e82587cdc36daa6275f6eec2c46909e8a071
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83976
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>