[BUGFIX] Make CE actions clickable after edit or visibility toggle
When a content element (CE) gets edited or the visibility is toggled,
the page module is reloaded with an anchor in the URL to scroll to the
content element that was updated.

After #103381 it was not possible anymore to click the "+ Content"
button above such CE. The reason is that a pseudo element
`:target:before` is generated with a specific height and margin to
overcome the height of the floating module docheader, overlaying the
CE action button.
In the patch mentioned before, `overflow: hidden` was removed to make
the `box-shadow` of the CE visible, finally causing this bug.

To solve this issue the pseudo element add it #97036 is replaced by the
CSS option scroll-padding-top. It incorperates the maximum docheader
size plus a slight margin that is slightly lower than the grid margin –
but not equal to – to move shadows of adjacent elements out of sight.

Resolves: #103672
Related: #103381
Related: #97036
Releases: main, 12.4
Change-Id: I8d5c9eadd45e5507e39b1fb728de65d31c1d6689
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83998
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Ayke Halder <mail@ayke-halder.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Ayke Halder <mail@ayke-halder.de>
Reviewed-by: Willi Wehmeier <wwwehmeier@gmail.com>
Tested-by: André Buchmann <andy.schliesser@gmail.com>
Tested-by: Willi Wehmeier <wwwehmeier@gmail.com>
Reviewed-by: André Buchmann <andy.schliesser@gmail.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>