[BUGFIX] Align search result output in "Configuration" backend module
As in the backend module "Configuration", search results can be
displayed as badges next to the search input field. When the search
input field is displayed with a label instead of placeholder text, the
badge was displayed out of place. This patch aligns the badge to the
search input field.

In addition, the search field now takes up full width for small
breakpoints.

Resolves: #103716
Releases: main
Change-Id: I322b4ab4c62823fa6dcf9d00fc30bc8671c57f09
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84013
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>