[BUGFIX] Omit wrong value on inserting record
When inserting a record via the element
browser, the value argument needs either
to be formatted manually or left empty,
allowing the element browser to format
the identifier automatically.

By omitting the value in the `PageBrowser`
module, the value is now correctly formatted
by the element browser to "pages_<uid>".

This makes the "node-action" usable for
TCA group fields, which allow more than
one record type.

The same logic is also used by the
`BrowseDatabase` module, which also
omits the value argument.

Resolves: #103700
Releases: main, 12.4
Change-Id: I80e2cc696d85b5ec83432c35a263b19f9752b0ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84553
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>