[TASK] Use light mode only as default for now
We're currently in a mixed mode where some components
support dark mode, while others don't. To facilitate the
migration of components currently incompatible with dark
mode, we're setting the default to light mode for now. This
allows us to gradually migrate components and enable the
dark mode switch once both a component and its context
are ready for it.

Resolves: #103723
Releases: main
Change-Id: I39e31d9b7170ac5592f58afe9f9dced0fd14c4c4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84020
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>