[TASK] Coloring the badges in "History/Undo" module
Since #103670 the "Workspaces" module has a slightly different look.
In order to keep the "History/Undo" module visually synchronous, the
badges are now displayed in color - depending on the action type.

In addition, the label for the action type "add" was changed based on
the wording from the "Workspaces" module.

Resolves: #103765
Releases: main
Change-Id: I7f63fc05fe766db8ee9fd85d12720d247009125f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84087
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Chris Müller <typo3@brotkrueml.dev>
Tested-by: Chris Müller <typo3@brotkrueml.dev>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>