[TASK] Update card layout and heading structure in "About" module
The "About" backend module provides overviews of the modules available
in the system, installed external libraries, authors of installed
extensions, as well as general information about TYPO3 and its
community.

This patch standardizes the appearance of the cards in this module with
those from the "Admin Tools" area. In addition, the heading structure
has been updated and some headings have been added.

In addition, links are now colored blue, similar to those in the
"System Information" toolbar element, and email links in the extension
authors list were removed because they violated the Content Security
Policy.

Resolves: #104009
Releases: main
Change-Id: Ia3a0811e973edea579e8d66d2f0d7f01e9ded0c2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84400
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>