[TASK] Make TS [END] an own include object
The new v12 TypoScript parser works pretty well. It is time
to rule out some of its rough edges in v13: In hinsight,
the biggest structural issues of the rewrite reside in the
IncludeTree structure, which has been oversimplified at some
places.

One problem is reflected within the BE modules 'TypoScript ->
Included TypoScript' and 'Page TSconfig -> Include page TSconfig',
which are not as easy to grasp as they could be. Those are
essentially a straight view of the created IncludeTree.

The patch changes a single decision within these trees:
"condition end" token ("[END]" / "[GLOBAL]") have been added
to the line stream of the currently given tree object until
now, but they should be a dedicated node reflected in the tree.

The patch turns "condition stop" into dedicated objects within
the tree. This makes BE include tree display more easy to grasp
and allows simplified scanning for missing and superfluous condition
stop lines in the future. It is also a base for better rendering
of the "code view" within the modules.

Additionally, condition and condition-else now render their
condition tokens in the BE modules, which should help
integrators to understand the tree more easily by simplified
matching to known parts of TS source.

Resolves: #103990
Related: #97816
Related: #102103
Releases: main
Change-Id: Ifb375ba321d98a17a372f49a5a096968387893ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84511
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>