[TASK] Introduce Alert web component
A new web component called `<typo3-backend-alert>`,
which is based on the Bootstrap alert component,
is introduced. It will be used to streamline the
use of this component through backend code.

The web component can therefore be used in
any JavaScript module or directly in HTML markup.

This is the first step towards unifying the alert,
notification, and callout components in the backend.
These components share a lot of common logic and
should therefore have a common ancestor.

Note: The new component might receive further
changes in the future and is therefore introduced
as internal for now.

Resolves: #103699
Releases: main
Change-Id: Ie31ef032ad1d3bb28553f518eff0a7673b53257d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84535
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>