[BUGFIX] Prevent backend errors for invalid backend_layout config
When the configuration for `mod.web_layout.BackendLayouts`
contains column definitions with an empty content, this
will make both the list view and the normal page view
in the backend fail.

This makes using the GUI editing of the 'backend_layout'
records impossible to be fixed/addressed.

This patch now does a more lax check on the
'columns.' array to prevent PHP error notices on
invalid array key access.

Resolves: #103425
Releases: main, 12.4, 11.5
Change-Id: I68c9b5a1ee8f0a70e3808338aa9dd2ed6f12c882
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84572
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>