[TASK] Remove conflicting card styles
This patch removes the Bootstrap "card" component as the
cards present in the Backend are implemented using the
classes and styling of TYPO3.

The conflict comes from a time in which Bootstrap (3.x)
only had panel-* classes and for the new card component
the classes were modelled after Bootstraps naming scheme.
This did not end up matching what Bootstrap implemented
finally, causing confusion within some developers when
using cards in the backend.

Resolves: #99233
Related: #94529
Releases: main
Change-Id: Ia1bf1f453310cc452059ec9396c2f54be87c2150
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76870
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>