[TASK] Use module title as link/button title in module menu
In the module menu, in addition to the visible module title, the short
description is inserted as a link/button title tag. The title attribute
is useful when the module menu is minimized. The short descriptions
previously used as titles are sometimes too long or not very helpful in
identifying the module. With this patch, the module titles are used as
link/button title tag.

Resolves: #104059
Releases: main
Change-Id: I19e8d9cebdbea4c6837373994c14954183069824
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84621
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>