[FEATURE] Improve backend notifications display and handling
Currently, the notification/alert container does not allow scrolling
when multiple notifications exceed the viewport height and
it is not possible to clear all displayed notifications.

This has been changed to allow scrolling within the
container. On top of this, it is now also possible
to focus the container using tab and scroll up and down
using the arrow keys and clear all notifications with a
single click.

Resolves: #104069
Releases: main
Change-Id: I6f4c27bff84942a3f161310da85a04957648218d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84659
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>