[BUGFIX] Properly handle checkbox state changes in FormEngine
Using the multi record selection, it's possible
to change the state of checkboxes via multiple
ways, e.g. by clicking on the table row.

The central functionality for changing the
connected checkbox state now also considers
the disabled attribute and therefore only
changes the checked state in case the
checkbox is not disabled.

Additionally, next to the custom event, which
is triggered on checkbox state changes, the
multi record selection now also dispatches
the standard "change" event. FormEngine
validation for example is listening on
this event to mark the field as changed.

Resolves: #104142
Releases: main, 12.4
Change-Id: I9854709ee624d88821757fff3baf38c892e5f11d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84823
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>