[TASK] Deprecate `Utility.updateQueryStringParameter()`
The method `Utility.updateQueryStringParameter()` from the module
`@typo3/backend/utility.js` was introduced in TYPO3 v8 as a bugfix for
highlighting in the old ExtJS-based page tree. With removal of ExtJS in
TYPO3 v9, the method became unused since then.

Since a safe removal of the method cannot be guaranteed as this point,
it is therefore deprecated.

Resolves: #104154
Related: #74180
Releases: main
Change-Id: I90f0ff4c997ab31d690ff6f518f38ffb186058dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84786
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>