[TASK] Update display of overview table in "Redirects" module
The redirects overview table has been revised to improve the UX.

Previously, the table fields took up as much space as their content
allowed. Redirects with a long source and/or destination path stretched
the table rows, which caused unwanted horizontal scrolling. The patch
now fixes this and sets a maximum width for these fields.

Furthermore, the "Status Code" and "Protected" fields are now also
managed as separate table columns. The record icon is now placed first
after the multi-record selection checkbox, as with other tables used in
the system.

Further actions, such as viewing the redirect and resetting the hit
counter, have been moved to a separate button group.

Resolves: #104166
Related: #93912
Releases: main
Change-Id: I5844e1e95ea9f3b1e9a2b06ca91dce9db3cf495e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84805
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>