[BUGFIX] Fix backend page tree filter with special characters
When filtering the page tree with characters that have
special meaning for a regular expression (for example,
")" or "(" or "?" and some others), the filtering would
return a "Page Tree error".

JavaScript does not have something like "preg_quote",
so this uses the escape string provided by the MDN
with fixed special characters.

Resolves: #104092
Releases: main, 12.4, 11.5
Change-Id: I4a29c62eaca8b7643c8ef78dda1d8337b3e1e53a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84820
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>