[BUGFIX] Fix invalid FormEngine HTML structure on edge-case
When the TCA type "group" is rendered in FormEngine, under
certain circumstances the HTML structure would close add
`<div>` elements without opening those first. This will
lead to a record not being editable, because variables are
wrongly transferred due to the browser-rendering of wrong
HTML nesting levels.

This behaviour is an edge case and only happens with this TCA
configuration:

* either `config[hideMoveIcons]` is set to `true`
* or `config[hideDeleteIcon]` is set to `true`

Resolves: #98436
Releases: main, 12.4, 11.5
Change-Id: Ibb28c6130051b3dae92b3285df0831e84634e210
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85012
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>