[BUGFIX] Respect "doSaveFieldName" in FormEngine again
It's possible to define a hidden input field
in FormEngine, which will be set to '1' when
the form is submitted, e.g. with a *save*
button. Therefore the "doSaveFieldName" option
of the `$resultArray` can be used.

This option is now respected again by
passing it to the FormEngine component.

Resolves: #104232
Releases: main, 13.4
Change-Id: I7737276486887db151b65d75fc85e6e50c6005bc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86629
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>