[TASK] Introduce CSS Color Tokens
With this patch, we are introducing color tokens
for use across the system. These tokens are provided
as CSS variables.

We generate the tokens by starting from a base value
and modifying the luminance, rather than shading or
tinting the current color. This approach ensures a
consistent color scheme.

Additionally, we are providing new surface variants
for the general states used in the backend, such as
primary, secondary, danger, and success. These will
be used in components and will back utility classes like
`.bg-<color>` and `.text-bg-<color>`.

Resolves: #104315
Releases: main
Change-Id: If39719eb23e1adf5ddb67c37380f47ed6627b968
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85132
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>