[BUGFIX] Respect condition state for PageTSConfig runtime cache
The method `BackendUtility::getPagesTSconfig()` has been using
a two level cache since #100701 to reduce the memory footprint.

Technically, a cache entry `pageTsConfig-pid-to-hash-<uid>` is
written pointing to a concrete PageTsConfig instance based on
relevant cache data. The goal was to reuse a `PageTsConfig`
instance when the config does not differ because it is inherited
from the rootline.

However, this approach missed the fact that even with the same
basic content, conditions can be evaluated differently for
distinct pages which can result in a different PageTsConfigArray.

This change adds a hash of the matched conditions to the second
cache level. As a consequence, the TS configuration has to be
evaluated at least once for each distinct page UID.

In other words, this change trades off runtime performance for
correctness.

It may be possible to investigate and find a broader approach to
improve performance by doing the condition evaluation independent
of the TS configuration parsing, but this is out of scope and needs
to be done in dedicated change.

Resolves: #104318
Related: #100701
Releases: main, 13.4
Change-Id: Id30183a0e312c121a90d8802c05d359acbe73278
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88968
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>