[TASK] Use explicit flatpickr types
The date time picker module `@typo3/backend/date-time-picker` uses
flatpickr, but wasn't aware of it due loose `any` usages. The code is
modified to make use of the typings shipped with flatpickr.

To solve this long-standing issue, the module identifier is changed from
`flatpickr/flatpickr.min` to `flatpickr`, which requires an additional
entry in the `JavaScriptModules.php` configuration, but allows to remove
a module declaration.

Resolves: #104338
Releases: main, 12.4
Change-Id: I3feed53627c639e21e4812855fd4ce2defe91c14
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85188
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>