[BUGFIX] Do not use getIconHtml() for other language thumbnail
The `OtherLanguageThumbnails` field wizard renders
thumbnails of each crop variant from the default
language for images in translations. However,
on the first view, the cropped (processed) image
is not created yet, leading to the public url
being an absolute url to the `image_processing`
route, e.g. "https://example.com/typo3/image/process?token=abc&id=123".

This URL can not be used by `getIconHtml()` to
create the markup. However the function call
is not needed at all, since we can just render
the markup directly as no icon is about to be
generated.

Resolves: #104355
Releases: main, 12.4
Change-Id: Ic624f8fd7114da18e01d7f33a9c8714bc8db026c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85221
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>