[TASK] Make text colors context aware
We replace the default bootstrap `text-<color>`
utility CSS classes with an own color-scheme
aware implementation.

In contrast to the original bootstrap color
utilities we ensure that the new set is
accessible and optimize the color contrast
for all values. This results in darker colors for
light-mode and lighter colors for dark-mode.

Resolves: #104362
Releases: main
Change-Id: I8e36394435aa523069aad784f54b632bff489eb8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85223
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>