[TASK] Avoid view->getTemplatePaths()
Fluid standalone AbstractTemplateView has
method getTemplatePaths() which is just
a wrapper around the same operation on
RenderingContext.

This method has been marked as deprecated
in Fluid standalone 2.14, and will be
removed in Fluid standalone v4.

The patch avoids some last usages.

Note this is a pre-patch to a bigger
patch that will change view handling in
extbase, which will come with proper
deprecation notes.

Resolves: #104446
Releases: main
Change-Id: I495ceac71cd998550fb930b45d8f34513e7cce3a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85318
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Praetorius <simon@praetorius.me>