[BUGFIX] Use localization parent for translated site root preview
In the edit view of a page, the page UID itself is used as the target
page for the preview button. This does not work for translated site
roots, as they cannot be mapped to a site configuration.

This uses the localization parent as the preview target if the current
page is a translated site root.

Resolves: #100406
Releases: main, 13.4
Change-Id: I589eb7c3a6b280257f9e2dbd610e3832402b5242
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88893
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>