[TASK] Modernize position map styling (a bit)
TYPO3 still uses the infamous position maps (lists with arrow icons) to
allow an user to re-arrange content elements. The position maps are used
in two places:

* "New Page Content" wizard in record list
* "Re-position content element" in record list

This commit aims to improve the styling of position maps and streamline
them with the page module, without the need of major refactorings.

Theoretically, TYPO3 has API to render a "page module"-alike grid
anywhere, but it is currently coupled too tight to page module behavior
and its needs.

Resolves: #104618
Releases: main
Change-Id: I6aa25e60e8ae95a88f6e73a86a910442bda722f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85624
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Ayke Halder <mail@ayke-halder.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Ayke Halder <mail@ayke-halder.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>