[TASK] Remove ancient server side 'no cookie' detection
There is ancient BE login related server side code to
detect if cookies can be set. This never kicks in and
is removed with the patch.

When 'setting cookies' is disallowed in a browser (see
instructions for firefox in issue), JS showCookieWarning()
detects this and enables an info box.
If JS is disabled as well, a default enabled info box
"Please enable JavaScript" is shown.

We're safe client side and can remove the server side
check that triggers a fishy reload for detection.

Resolves: #104742
Releases: main
Change-Id: Id2585f77f097fd538766053d2b79cc40497737f6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85787
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Simon Praetorius <simon@praetorius.me>