[BUGFIX] Prevent undefined array key exception in FlexFormValueFormatter
The `getProcessedElements` function in `FlexFormValueFormatter`
throws an undefined array key exception, when the given
dataStructure contains a field with a relation table having a
`label_userFunc` and when the given valueStructure does not contain
a value for the field.

This change adds a fallback to an empty string, when
`$valueStructure[$elementKey]['vDEF']` is not set.

Resolves: #104760
Releases: main, 12.4
Change-Id: I133028b341d6df10a32f56858f8be8cfe951625a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85765
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>