[BUGFIX] Allow login notification mail with f:translate VH
With #100307 the authentication system got streamlined and
prepared to pass the request and added on dispatched PSR-14
events. To mitigate breaking change nullable method parameter
signature has been used along with keeping the fallback
handling retrieving the global request or creating a valid
request from globales within the `EmailLoginNotification`
event listener.

The code has been additionally adjusted for issues #100128
and #99864 in case of MFA authentication but missed to pass
the request correctly to the added intermediate method.

This change passes the request down to the event dispatching
method which contains the application type request argument.
To mitigate issues in systems using xclasses or test setups
the fallback handling in `EmailLoginNotification` is modified
to set the BE application type in case a request is created
from globals.

Both solution parts ensures to have a proper application
type set and that `f:translate` viewhelper can be used in
system mail templates.

Resolves: #104809
Related: #100128
Related: #99864
Related: #100307
Releases: main, 12.4
Change-Id: Ib23a035bfed38c7dc03db111e42f6e98b12e3a4b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85862
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>